CytoPipeline class

Class representing a flow cytometry pipeline, and composed of two processing queues, i.e. lists of CytoProcessingStep objects :

a list of CytoProcessingStep(s) for pre-calculation of scale transformations per channel

a list of CytoProcessingStep(s) for the pre-processing of flow frames

# S4 method for CytoPipeline
show(object)

# S4 method for missing
CytoPipeline(
  object,
  experimentName = "default_experiment",
  sampleFiles = character()
)

# S4 method for list
CytoPipeline(object)

# S4 method for character
CytoPipeline(object)

# S3 method for CytoPipeline
as.list(x, ...)

experimentName(x)

experimentName(x) <- value

sampleFiles(x)

sampleFiles(x) <- value

Arguments

object

a character() containing a JSON input

experimentName

the experiment name

sampleFiles

the sample files

x

a CytoPipeline object

...

additional arguments (not used here)

value

the new value to be assigned

Value

for as.list.CytoPipeline: the obtained list

Slots

scaleTransformProcessingQueue

A list() of CytoProcessingStep objects containing the steps for obtaining the scale transformations per channel

flowFramesPreProcessingQueue

A list() of CytoProcessingStep objects containing the steps for pre-processing of the samples flow frames

experimentName

A character() containing the experiment (run) name

sampleFiles

A character() vector storing all fcs files to be run into the pipeline

saveScaleTransform

if TRUE, will save rds object storing scale

scaleTransformFile

basename of the file to use to save the scale