perform QC with flowAI

this function is a wrapper around flowAI::flow_auto_qc() function. It also pre-selects the channels to be handled (=> all signal channels)

qualityControlFlowAI(
  ff,
  preTransform = FALSE,
  transList = NULL,
  outputDiagnostic = FALSE,
  outputDir = NULL,
  ...
)

Arguments

ff

a flowCore::flowFrame

preTransform

if TRUE, apply the transList scale transform prior to running the gating algorithm

transList

applied in conjunction with preTransform

outputDiagnostic

if TRUE, stores diagnostic files generated by flowAI in outputDir directory

outputDir

used in conjunction with outputDiagnostic

...

additional parameters passed to flowAI::flow_auto_qc()

Value

a flowCore::flowFrame with removed low quality events from the input

Examples

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

suppressWarnings(ff_QualityControl <- 
    qualityControlFlowAI(fsRaw[[2]],
                         remove_from = "all", # all default
                         second_fractionFR = 0.1,
                         deviationFR = "MAD",
                         alphaFR = 0.01,
                         decompFR = TRUE,
                         outlier_binsFS = FALSE,
                         pen_valueFS = 500,
                         max_cptFS = 3,
                         sideFM = "both",
                         neg_valuesFM = 1))
#> Applying flowAI method...
#> Quality control for the file: sample_Donor2
#> 4.52% of anomalous cells detected in the flow rate check. 
#> 0% of anomalous cells detected in signal acquisition check. 
#> 0.1% of anomalous cells detected in the dynamic range check.