perform QC with flowCut

this function is a wrapper around flowCut::flowCut() function. It also pre-selects the channels to be handled (=> all signal channels)

qualityControlFlowCut(
  ff,
  preTransform = FALSE,
  transList = NULL,
  outputDiagnostic = FALSE,
  outputDir = NULL,
  verbose = TRUE,
  ...
)

Arguments

ff

a flowCore::flowFrame

preTransform

if TRUE, apply the transList scale transform prior to running the gating algorithm

transList

applied in conjunction with preTransform

outputDiagnostic

if TRUE, stores diagnostic files generated by flowCut in outputDir directory

outputDir

used in conjunction with outputDiagnostic

verbose

if TRUE messages comments on the QC process

...

additional parameters passed to flowCut::flowCut()

Value

a flowCore::flowFrame with removed low quality events from the input

Examples

rawDataDir <-
    paste0(system.file("extdata", package = "CytoPipeline"), "/")
sampleFiles <-
    paste0(rawDataDir, list.files(rawDataDir, pattern = "sample_"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

ff_QualityControl <-
    qualityControlFlowCut(
        fsRaw[[2]],
        MaxContin = 0.1,
        MeanOfMeans = 0.13,
        MaxOfMeans = 0.15,
        MaxValleyHgt = 0.1,
        MaxPercCut = 0.3,
        LowDensityRemoval = 0.1,
        RemoveMultiSD = 7,
        AlwaysClean = FALSE,
        IgnoreMonotonic = FALSE,
        MonotonicFix = NULL,
        Measures = c(1:8))
#> Applying flowCut method...
#> The FileID is: 2022_12_19__14_38_29
#> The flow frame is not monotonically increasing in time.
#> None deleted from flowCut low dens removal.
#> Removing segments 2 out of 10 segments.
#> The file has been flagged. The largest continuous jump was larger than 10% of the range of the 2-98 percentile of the full data.
#> Type of Gating: MaxPercCut.
#> The file has been flagged FFTT
#> Cleaning completed in: 00:00:00
#> Running flowCut a second time.
#> The flow frame is not monotonically increasing in time.
#> None deleted from flowCut low dens removal.
#> None deleted from flowCut segment removal.
#> The file has been flagged. The largest continuous jump was larger than 10% of the range of the 2-98 percentile of the full data.
#> Type of Gating: MaxPercCut.
#> The file has been flagged FFTT
#> Cleaning completed in: 00:00:00